+++ sentwits/sentwits.py
@@ -1,40 +1,58 @@
+import pandas as pd
 import requests
-import collections
 
 class COMMENTS:
 
-    def __init__(self,ticker,timeframe,limit):
+    def __init__(self, ticker, limit, timeframe='1-month'):
         self.ticker = ticker
         self.timeframe = timeframe
         self.limit = limit
 
-    def stocktwits_comment_sentiment(self):
+    def GetComments(self):
         if self.limit > 30:
             limit = 30
 
-        sentiment = []
+        userid = []
+        body = []
+
+        url = requests.get('https://api.stocktwits.com/api/2/streams/symbol/{}.json?filter=all&limit={}'.format(self.ticker, self.limit))
+        api = url.json()
+
+        for x in range(self.limit):
+            body.append(api['messages'][x]['body'])
+            userid.append(api['messages'][x]['user']['username'])
+
+        return pd.DataFrame({'Username': userid, 'Message': body})
+
+    def GetSentiment(self):
+        if self.limit > 30:
+            limit = 30
+
+        bull = []
+        bear = []
 
         url = requests.get(
             'https://api.stocktwits.com/api/2/streams/symbol/{}.json?filter=all&limit={}'.format(self.ticker, self.limit))
         api = url.json()
 
-        for x in range(limit):
-            i = ''
+        for x in range(self.limit):
             if api['messages'][x]['entities']['sentiment'] is not None:
                 i = api['messages'][x]['entities']['sentiment']['basic']
-                sentiment.append(i)
+                if i == 'Bullish':
+                    bull.append(i)
+                if i == 'Bearish':
+                    bear.append(i)
 
-            print(api['messages'][x]['user']['username'] + ':', api['messages'][x]['body'], "\033[1m" + i + "\033[0m")
-            print('')
-        return 'Bullish:', collections.Counter(sentiment)['Bullish'], 'Bearish:', collections.Counter(sentiment)[
-            'Bearish']
+        df = pd.DataFrame.from_dict({'Bullish': [len(bull)],
+                                     'Bearish': [len(bear)]})
+        return df
 
 class SENTIMENT:
+
     def __init__(self):
         pass
 
-
-    def stocktwits_sentiment_ratio(self=True):
+    def GetSentimentRatio(self=True):
         tickers = []
 
         # Find all trending stocks today
@@ -71,27 +89,51 @@ class SENTIMENT:
             indicator = round((bull / bear), 3)
             return indicator
 
+class USER:
 
-        # Returns top watched tickers
+    def __init__(self, userid, limit):
+        self.userid = userid
+        self.limit = limit
 
-        # Returns top watched tickers
+    def GetComments(self):
 
-    def stocktwits_top_watched(self=True):
+        body = []
+        times = []
+        url = requests.get('https://api.stocktwits.com/api/2/streams/user/{}.json?filter=all&limit={}'.format(self.userid, self.limit))
+        api = url.json()
+        for x in range(self.limit):
+            body.append(api['messages'][x]['body'])
+            times.append(api['messages'][x]['created_at'])
+
+        return pd.DataFrame({'Message': body, 'Datetime': times})
+
+    def GetFollow(self):
+        url = requests.get('https://api.stocktwits.com/api/2/streams/user/{}.json?filter=all&limit={}'.format(self.userid, self.limit))
+        api = url.json()
+        followers = api['user']['followers']
+        following = api['user']['following']
+
+        return pd.DataFrame({'Followers': [followers], 'Following': [following]})
+
+class GENERAL:
+    def __init__(self):
+        pass
+
+    def GetTopWatched(self=True):
         tickers = []
         url = requests.get('https://api.stocktwits.com/api/2/watchlists/top_watched.json')
         api = url.json()
 
         for x in api['top_watched']:
             tickers.append(x['symbol'])
-        return tickers
+        return pd.DataFrame({'Tickers': tickers})
 
         # Returns list of trending tickers today
 
-    def stocktwits_trending(self=True):
+    def GetTrending(self=True):
         tickers = []
         url = requests.get('https://api.stocktwits.com/api/2/trending/symbols.json')
         api = url.json()
         for x in api['symbols']:
             tickers.append(x['symbol'])
-        return tickers
-
+        return pd.DataFrame({'Tickers': tickers})