+++ CHANGELOG.rst
@@ -1,3 +1,12 @@
+0.0.3
+-------
+- Change of function names, 'GetSentiment' now return a dataframe.
+- Added a USER class.
+- Added af GENERAL class
+- Most datasets now return a DataFrame
+- Fixed some bugs.
+
+
 0.0.2
 -------
 - Small fixes.